Header fixed in all workshop pages plus booking status has been chang…

+++ database/migrations/2018_02_26_102910_create_billings_table.php
@@ -24,7 +24,7 @@ class CreateBillingsTable extends Migration
             $table->decimal('amount', 11, 2);            
             $table->decimal('paid_amount', 11, 2)->nullable();
             $table->decimal('lead_charges',7,2);
-            $table->tinyInteger('ratings')->nullable();    
+            $table->decimal('ratings', 2,1)->nullable();    
             $table->text('review')->nullable();
             $table->timestamps();
         });
